idc: extract common protocol functions
Extracts IDC protocol related code to common file.
Platform specific driver has just the implementations
for hardware specific operations.
